refactor(cache): extract ImageCache + pure budget helper
Move ImageCache, ImageCacheDelegate, and Logger.imageCache out of
PreviewView.swift into their own file. Promote the fullRes budget
computation to a pure ImageCacheBudget.compute(physicalMemory:aggressive:)
free helper so tests can exercise the full RAM range without standing
up a real machine. Behaviour at runtime is unchanged - same singletons,
same call sites; AdvancedTab and CullingConfig now resolve physical
memory and aggressive flag at the call site instead of relying on the
old static method.

Adds ImageCacheBudgetTests covering the floor, balanced/aggressive
fractions, and the [minBytes, maxBytes] clamp.
